+++ app/models/LineIndex.scala
@@ -1,3 +1,3 @@
-package readme
+package models
 
 case class LineIndex(indentation: Int, title: String)
