+++ app/models/ReadmeForm.scala
@@ -1,3 +1,3 @@
-package readme
+package models
 
 case class ReadmeForm(content: String, githubUrl: String)
