+++ app/models/TOCResult.scala
@@ -1,3 +1,3 @@
-package readme
+package models
 
 case class TOCResult(table: Seq[String], fullText: String)
